SYMMETRICDS-562 - Fix DataService bug. Max data id that is currently handled in 2.x is Integer.MAX_VALUE. It should go all the way up to Long.MAX_VALUE

+++ symmetric-core/src/main/java/org/jumpmind/symmetric/route/DataRefGapDetector.java
@@ -73,7 +73,7 @@ public class DataRefGapDetector implements IDataToRouteGapDetector {
         long ts = System.currentTimeMillis();
         final int dataIdIncrementBy = parameterService.getInt(ParameterConstants.DATA_ID_INCREMENT_BY);
         long lastDataId = (Long) jdbcTemplate.query(sqlProvider.getSql("selectDistinctDataIdFromDataEventSql"),
-                new Object[] { ref.getRefDataId() }, new int[] { Types.INTEGER },
+                new Object[] { ref.getRefDataId() }, new int[] { Types.NUMERIC },
                 new ResultSetExtractor<Long>() {
                     public Long extractData(ResultSet rs) throws SQLException, DataAccessException {
                         long lastDataId = ref.getRefDataId();
